Move pydust_demo tests to pzspec/ folder
- Move test_pydust_mathlib.py to pzspec/test_mathlib.py
- Simplify tests by assuming pydust is available
- Auto-copy ffi.h in build.py (no manual step needed)
- Update README with simplified setup instructions
