+++ usage/pydust_demo/.gitignore
@@ -14,4 +14,4 @@ __pycache__/
 .venv/
 
 # PZSpec snapshots
-pzspec/
+pzspec/__snapshots__/
