+++ usage/pydust_demo/pzspec/__init__.py
@@ -1,0 +1,1 @@
+# PZSpec tests for pydust_demo
